third_party/blink/web_tests/fast/forms/label/labels-remove-htmlFor-label-expected.txt

Test that the labels list of a form control is dynamically updated when removing the htmlFor attribute that pointed to that form control.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check button
PASS element.labels.length is 1
PASS element.labels.length is 0
Check input
PASS element.labels.length is 1
PASS element.labels.length is 0
Check meter
PASS element.labels.length is 1
PASS element.labels.length is 0
Check output
PASS element.labels.length is 1
PASS element.labels.length is 0
Check progress
PASS element.labels.length is 1
PASS element.labels.length is 0
Check select
PASS element.labels.length is 1
PASS element.labels.length is 0
Check textarea
PASS element.labels.length is 1
PASS element.labels.length is 0
Check buttonType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check checkboxType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check colorType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check dateType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check datetimeType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check datetime-localType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check emailType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check fileType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check imageType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check monthType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check numberType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check passwordType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check radioType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check rangeType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check resetType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check searchType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check submitType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check telType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check textType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check timeType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check urlType
PASS element.labels.length is 1
PASS element.labels.length is 0
Check weekType
PASS element.labels.length is 1
PASS element.labels.length is 0
PASS successfullyParsed is true

TEST COMPLETE
